requirements.txt
flask
cassandra-driver[cloud]
python-dotenv
